refactor(common): Price, DisplayedName, Profanities 공통 모듈로 이동
