+++ src/main/java/kitchenpos/common/infra/Profanities.java
@@ -1,4 +1,4 @@
-package kitchenpos.products.infra;
+package kitchenpos.common.infra;
 
 public interface Profanities {
     boolean containsProfanity(String text);
