+++ src/main/java/kitchenpos/common/vo/DisplayedName.java
@@ -1,22 +1,24 @@
-package kitchenpos.menus.tobe.domain;
+package kitchenpos.common.vo;
+
+import kitchenpos.common.infra.Profanities;
 
 import javax.persistence.Column;
 import javax.persistence.Embeddable;
 import java.util.Objects;
 
 @Embeddable
-class DisplayedName {
+public class DisplayedName {
     private static final String EMPTY_NAME_MESSAGE = "이름은 비어있을 수 없습니다.";
     private static final String CONTAIN_PROFANITY_MESSAGE = "이름에 비속어가 포함될 수 없습니다.";
 
     @Column(name = "name", nullable = false)
     private final String name;
 
-    DisplayedName(final String name, final Profanity profanity) {
+    public DisplayedName(final String name, final Profanities profanities) {
         if (null == name || name.isBlank()) {
             throw new IllegalArgumentException(EMPTY_NAME_MESSAGE);
         }
-        if (profanity.isContains(name)) {
+        if (profanities.containsProfanity(name)) {
             throw new IllegalArgumentException(CONTAIN_PROFANITY_MESSAGE);
         }
         this.name = name;
@@ -24,6 +26,10 @@ class DisplayedName {
 
     protected DisplayedName() {
         this.name = null;
+    }
+
+    public String getName() {
+        return name;
     }
 
     @Override
