--- src/main/java/kitchenpos/menus/tobe/domain/Profanity.java
@@ -1,5 +1,0 @@
-package kitchenpos.menus.tobe.domain;
-
-interface Profanity {
-    boolean isContains(String name);
-}
