+++ src/test/java/kitchenpos/menus/tobe/domain/DisplayedNameTest.java
@@ -1,5 +1,7 @@
 package kitchenpos.menus.tobe.domain;
 
+import kitchenpos.common.FakeProfanity;
+import kitchenpos.common.vo.DisplayedName;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
