--- src/test/java/kitchenpos/menus/tobe/domain/FakeProfanity.java
@@ -1,12 +1,0 @@
-package kitchenpos.menus.tobe.domain;
-
-import java.util.List;
-
-class FakeProfanity implements Profanity {
-    private final List<String> values = List.of("damn");
-
-    @Override
-    public boolean isContains(final String name) {
-        return values.contains(name);
-    }
-}
