+++ src/test/java/kitchenpos/products/application/FakePurgomalumClient.java
@@ -1,6 +1,6 @@
 package kitchenpos.products.application;
 
-import kitchenpos.products.infra.Profanities;
+import kitchenpos.common.infra.Profanities;
 
 import java.util.Arrays;
 import java.util.List;
@@ -15,6 +15,6 @@ public class FakePurgomalumClient implements Profanities {
     @Override
     public boolean containsProfanity(final String text) {
         return profanities.stream()
-            .anyMatch(profanity -> text.contains(profanity));
+            .anyMatch(text::contains);
     }
 }
